Why: Keep Track of Packages.
*

I'm sure there is a better way to do this, but I
figure this is a good way to keep track of the awesome
packages available at least on my Arch linux builds.

Later I plan to look at stow, dotbot, etc., for better
management.

This change addresses the need by: luciddeveloper

*