Why: It had to happen.. lol
*
So I have added a lot of changes of the last couple of days.
Due to overrall tiredness and busy-ness, I was unable to upload.
The newest changes is the revamping of egun-theme. I now use
kawagana.nvim theme which purports to be a combination of gruvbox
and tokyonights. It is delicious. There are several other changes.
Lots to come.

Next is one that I am pretty excited about but have not yet had the
cycles to work on is the CompileAndExecute function that I found.
Forgive me, I forgot your name, however, it is in the ref somewhere.
I plan to port it from vim to lua soon.

This change addresses the need by: luciddeveloper

*